Add a two-player mode that makes use of many mouse.
Add a "PlayerMode" setting to state.xml.
This option can be changed by typing "onemouse" or "twomice".
The first mouse to move controls the Yin and the second controls the Yang.
Mouse buttons and scrolling with manymouse are not yet implemented.

Currently there is a bug causing interference between the two controllers.