Solve the delayed manymouse events bug (I hope).
When manymouse used XInput2, when flushing the event queue, sometimes some of the events were not flushed, and only release later.
This caused the ball to sometimes "shoot" when returning from the pause menu.
I managed to reproduce this behavior without using enigma, so I know the problem is with manymouse.
I tried to work around this by by waiting for a while and then flushing again, and it seemed that it only reduced the chances of the bug happening, so probably the wait was too short.

This time I tried a change inside manymouse. I added a call for XSync() before each check for an event.
I don't really know what I am doing, but this seems to work.